requirements.txt
# Core Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
alembic>=1.7.0

# Database Drivers
psycopg2-binary>=2.9.0
redis>=4.0.0

# LangChain & AI 
langchain>=0.1.0
langchain-experimental>=0.1.0
langgraph>=0.1.0
openai>=1.0.0

# Vector Store 
pinecone-client>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Google Cloud
google-cloud-storage>=2.0.0
google-cloud-aiplatform>=1.0.0
google-cloud-pubsub>=2.0.0
google-cloud-bigquery>=3.0.0
google-cloud-logging>=3.0.0

# Security
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
bcrypt>=3.2.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.16.0
pytest-cov>=3.0.0
httpx>=0.20.0

# Utilities
python-json-logger>=2.0.0
tenacity>=8.0.0
structlog>=21.1.0
aiohttp>=3.8.0
cachetools>=4.2.0
